Revert "Extending code and renaming functions"
This reverts commit 456d7447e7d7fa84546238d3e72e1c7a017b37a3.
